Refactor Gender to inner enum in StudentProfileAttributes
