+++ src/main/java/teammates/ui/webapi/action/PutStudentProfileAction.java
@@ -2,7 +2,6 @@ package teammates.ui.webapi.action;
 
 import org.apache.http.HttpStatus;
 
-import teammates.common.datatransfer.attributes.Gender;
 import teammates.common.datatransfer.attributes.StudentProfileAttributes;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.exception.UnauthorizedAccessException;
@@ -55,7 +54,8 @@ public class PutStudentProfileAction extends Action {
         if ("".equals(editedProfile.nationality)) {
             editedProfile.nationality = getRequestParamValue("existingNationality");
         }
-        editedProfile.gender = Gender.getGenderEnumValue(getRequestParamValue(Const.ParamsNames.STUDENT_GENDER));
+        editedProfile.gender = StudentProfileAttributes.Gender.getGenderEnumValue(
+                                                                getRequestParamValue(Const.ParamsNames.STUDENT_GENDER));
         editedProfile.moreInfo = getRequestParamValue(Const.ParamsNames.STUDENT_PROFILE_MOREINFO);
         editedProfile.pictureKey = "";
